Extract CreateAnnuityDialogView from CreateAnnuityDialog
Split the 562-line CreateAnnuityDialog.tsx into a hook container
(CreateAnnuityDialog.tsx, 215 lines) plus a presentational
CreateAnnuityDialogView.tsx (357 lines), mirroring the existing
EditAnnuityDialog / EditAnnuityDialogView pattern. No behavior change; the
view's spec and story are re-pointed at the new file, and the barrel re-exports
CreateAnnuityDialogView (and its props) from the new module.

This lands ahead of the ESLint max-lines cap (PR #361, stacked on this branch)
so that adopting the whole-repo 400-line cap does not fail on this file.
