+++ src/components/annuities/index.ts
@@ -6,10 +6,9 @@ export { AnnuityListView } from "./AnnuityListView";
 export type { AnnuityPaymentHistoryTableProps } from "./AnnuityPaymentHistoryTable";
 export { AnnuityPaymentHistoryTable } from "./AnnuityPaymentHistoryTable";
 export type { CreateAnnuityInput } from "./CreateAnnuityDialog";
-export {
-  CreateAnnuityDialog,
-  CreateAnnuityDialogView,
-} from "./CreateAnnuityDialog";
+export { CreateAnnuityDialog } from "./CreateAnnuityDialog";
+export type { CreateAnnuityDialogViewProps } from "./CreateAnnuityDialogView";
+export { CreateAnnuityDialogView } from "./CreateAnnuityDialogView";
 export type { DeleteAnnuityDialogProps } from "./DeleteAnnuityDialog";
 export { DeleteAnnuityDialog } from "./DeleteAnnuityDialog";
 export type {
